Create ai-photo-realism.py
Turns an ai generated photo file into a file that is undetectable when someone uses code to detect ai generated photos. Note that human detectors could still have the opinion that it is ai generated but they will have little to no proof. Also the image is not changed for the human eye before and after running the script :)